[BRMP-6742] debug scripts

+++ scripts/init-migrations.sh
@@ -21,21 +21,29 @@ else
         formatted_print "ERROR: No Argument Provided" $SCRIPT_NAME
     else
         formatted_print "--> Initializing Django Migrations" $SCRIPT_NAME
+
+        # LOAD IN ENVIRONMENT VARIABLES
         if [ "$1" == "local" ]
         then
-            formatted_print '--> Setting Environment Variables' $SCRIPT_NAME
+            formatted_print '--> Reading Environment Variables' $SCRIPT_NAME
             if [ -f "$SCRIPT_DIR/../env/local.env" ]
             then
+                formatted_print '--> Importing Environment Variables' $SCRIPT_NAME
                 set -o allexport
                 source $SCRIPT_DIR/../env/local.env
                 set +o allexport
+                echo $SECRET_KEY
+                echo $AWS_BUCKET_NAME
             fi
         fi
+
+        # NAVIGATE TO FOLDER CONTAINING MANAGE.PY
         if [ "$1" == "local" ] || [ "$1" == "container" ]
         then
             APP_DIR="$SCRIPT_DIR/../sinwebapp"
         fi
 
+        # CACHE CUSTOM MIGRATIONS IN /DB/ FOLDER
         formatted_print '--> Copying Custom Data Migrations Into \e[4m/db/\e[0m Directory Before Scrubbing The Application' $SCRIPT_NAME
         if [ -f "$APP_DIR/api/migrations/0002_api_data.py" ]
         then
@@ -53,17 +61,33 @@ else
             cp -R "$APP_DIR/authentication/migrations/0002_authentication_data.py" "$APP_DIR/db/0002_authentication_data.py"
         fi
         
+        # CLEAN MIGRATIONS
         formatted_print "--> Cleaning Migrations In All \e[4m$APP_DIR/\e[0m Modules" $SCRIPT_NAME
-        formatted_print '--> Cleaning \e[4m/api/migrations/\e[0m Directory' $SCRIPT_NAME
-        rm -r $APP_DIR/api/migrations/
-        formatted_print '--> Cleaning \e[4m/api/authentication/\e[0m Directory' $SCRIPT_NAME
-        rm -r $APP_DIR/authentication/migrations/
+        if [ -d "$APP_DIR/api/migrations/" ]
+        then
+            formatted_print '--> Cleaning \e[4m/api/migrations/\e[0m Directory' $SCRIPT_NAME
+            rm -r $APP_DIR/api/migrations/
+        fi
+        if [ -d "$APP_DIR/authenication/migrations/" ]
+        then
+            formatted_print '--> Cleaning \e[4m/authenication/migrations/\e[0m Directory' $SCRIPT_NAME
+            rm -r $APP_DIR/authentication/migrations/
+        fi
 
+        # CREATE FRESH MIGRAITONS
         formatted_print "--> Creating New Model Migrations"
         cd $APP_DIR
         python manage.py makemigrations api --name init
         python manage.py makemigrations authentication --name init
         python manage.py makemigrations
+
+        # COPY CUSTOM MIGRATIONS BACK INTO APPLICAITON
+        if [ ! -d "$APP_DIR/api/migrations/" ]
+        then
+            formatted_print '--> No \e[4m/api/migrations/\e[0m Directory Detected, Creating New Directory' $SCRIPT_NAME
+            mkdir -p "$APP_DIR/authentication/migrations/"
+            touch "$APP_DIR/authentication/migrations/__init__.py"
+        fi
 
         formatted_print '--> Copying Custom Data Migrations Back Into Application' $SCRIPT_NAME
         cp -R "$APP_DIR/db/0002_api_data.py" "$APP_DIR/api/migrations/0002_api_data.py"
